fixed taxonomy editing not checking user roles

+++ app/routes/taxonomies.js
@@ -3,8 +3,25 @@
 var logger = require("../util/logger");
 var db = require('../sqlClient/mysqlclient').dbClient;
 
-exports.updateableTaxonomies = function(req, res){
+function isTaxonomyAdmin(req, res) {
+    var authToken = req.headers.authorization.split('Bearer ')[1];
 
+    var requestingUser = req.user.username;
+
+    return db.loadUserRoles(requestingUser).then(function (roles) {
+        var isTaxonomyAdmin = q.defer();
+        roles.forEach(function (role) {
+            if (role === 'TaxAdmin') {
+                isTaxonomyAdmin.resolve(true);
+                return isTaxonomyAdmin.promise;
+            }
+        }, genericError);
+        isTaxonomyAdmin.resolve(false);
+        return isTaxonomyAdmin.promise;
+    });
+}
+
+exports.updateableTaxonomies = function(req, res){
     db.updateableTaxonomies().then(function (updateableTaxonomies) {
         res.status(200).send(updateableTaxonomies);
     }, function (err) {
@@ -14,21 +31,35 @@ exports.updateableTaxonomies = function(req, res){
 };
 
 exports.addTaxonomyElement = function(req, res){
-    var body = req.body;
 
-    db.addTaxonomyElement(req.params.taxonomy, body).then(function () {
-        res.status(200).end();
-    }, function (err) {
-        logger.log('Error while retrieving adding new taxonomy element.');
-        res.status(501).send('Error while retrieving adding new taxonomy element.');
+    isTaxonomyAdmin(req, res).then(function (isTaxonomyAdmin) {
+        if (isTaxonomyAdmin) {
+            var body = req.body;
+
+            db.addTaxonomyElement(req.params.taxonomy, body).then(function () {
+                res.status(200).end();
+            }, function (err) {
+                logger.log('Error while retrieving adding new taxonomy element.');
+                res.status(501).send('Error while retrieving adding new taxonomy element.');
+            });
+        } else {
+            res.status(403).send('Only taxonomy admins can edit taxonomies.');
+        }
     });
 };
 
 exports.deleteTaxonomyElement = function(req, res){
-    db.deleteTaxonomyElement(req.params.taxonomy, req.params.name).then(function () {
-        res.status(200).end();
-    }, function (err) {
-        logger.log('Error while retrieving deleting taxonomy element.');
-        res.status(501).send('Error while retrieving deleting taxonomy element.');
+
+    isTaxonomyAdmin(req, res).then(function (isTaxonomyAdmin) {
+        if (isTaxonomyAdmin) {
+            db.deleteTaxonomyElement(req.params.taxonomy, req.params.name).then(function () {
+                res.status(200).end();
+            }, function (err) {
+                logger.log('Error while retrieving deleting taxonomy element.');
+                res.status(501).send('Error while retrieving deleting taxonomy element.');
+            });
+        } else {
+            res.status(403).send('Only taxonomy admins can edit taxonomies.');
+        }
     });
 };
